+++ src/index.ts
@@ -1,7 +1,9 @@
 export { default as ConfigProvider } from "./configProvider/ConfigProvider";
-export { defaultConfig } from "./configProvider/ConfigProvider";
+export { default as defaultConfig } from "./configProvider/defaultConfig";
 
 export * from "./common/headlessService/types";
+
+export { default as Link } from "./link/Link";
 
 export { default as Navigation } from "./navigation/Navigation";